// packages/frontend/src/generated/types/CreatePhysicalTransferRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { PhysicalTransferItemCompositeRequest } from './PhysicalTransferItemCompositeRequest.ts'

export interface CreatePhysicalTransferRequest {
  /**
   * @type string, date-time
   */
  date: string
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string, date-time
   */
  endCargoOps: string
  /**
   * @type array
   */
  items: PhysicalTransferItemCompositeRequest[]
  /**
   * @type string, date-time
   */
  startCargoOps: string
}