// packages/frontend/src/generated/types/PhysicalTransferItemCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface PhysicalTransferItemCompositeRequest {
  /**
   * @type string
   */
  amount: string
  /**
   * @type string, uuid
   */
  contractorId: string
  /**
   * @type string, uuid
   */
  fromStorageId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string, uuid
   */
  toStorageId: string
}