// packages/frontend/src/generated/zod/createPhysicalTransferRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { CreatePhysicalTransferRequest } from '../types/CreatePhysicalTransferRequest.ts'
import * as z from 'zod'
import { physicalTransferItemCompositeRequestSchema } from './physicalTransferItemCompositeRequestSchema.ts'

export const createPhysicalTransferRequestSchema = z.object({
  date: z.string().datetime(),
  documentNumber: z.string(),
  endCargoOps: z.string().datetime(),
  items: z.array(z.lazy(() => physicalTransferItemCompositeRequestSchema)),
  startCargoOps: z.string().datetime(),
}) as unknown as ToZod<CreatePhysicalTransferRequest>