// packages/frontend/src/generated/zod/physicalTransferItemCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { PhysicalTransferItemCompositeRequest } from '../types/PhysicalTransferItemCompositeRequest.ts'
import * as z from 'zod'

export const physicalTransferItemCompositeRequestSchema = z.object({
  amount: z.string(),
  contractorId: z.string().uuid(),
  fromStorageId: z.string().uuid(),
  productId: z.string().uuid(),
  toStorageId: z.string().uuid(),
}) as unknown as ToZod<PhysicalTransferItemCompositeRequest>